+++ tests/functional/spec/features/calculated_summary/calculated_summary.spec.js
